requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Baker==1.3
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
exceptiongroup==1.1.1
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
kombu==5.2.4
oauthlib==3.2.2
packaging==23.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
